Dev/2.1.0 map distribution and update of Figures (#278)
Map Distribution:
A chapter on Map Distribution was added. In comparison to the previously introduced concept are mapIds describing a certain space in the working area of the robot that stays constant. Updates to the maps are done by using the mandatory mapVersion to work in changes to the environment. Functionalities to trigger the download, the deletion, the activation were added and an update to the vehicle state was introduced.

Figures:

Multiple figures were updated in the style of the Proposed Figure 8 in PR Feature/update figure 8 #265. The style was adjusted and the markdown files were added to ensure easy tracking and adaption of existing Figures.
To ensure a single point of truth, all figure names were changed from FigureX.png to a describing naming of the Figure. Now changes in the Figure Sequence within the documents do not lead to changes in the naming of the Figures.
The overall order and numbering of the Figures was adjusted as well.


---
* Added Chapter on Maps
* Added vehicle status changes
* Added mapVersion on multiple points, added texts.
* Updated ToC
* complete restructuring of the maps distribution chapter, added instantActions, status update and descriptions.
* Updating the wording of the map distribution, adding image for correct use across multiple maps.
* Updated map distribution: making mapVersion not optional anymore, changing map status active to enabled, allowing multiple maps to be enabled at the same time.
* Changes from VDA 5050 Meeting December 2023: - Map Version is mandatory to identify maps uniquely. This is necessary, since the mapId is a constant identification of certain areas of the vehicles workspace due to the LIF specification.
- It is no longer possible to enable maps by using them in an order. The master control has to ensure, that all the necessary versions are present on the vehicle.
- In the order, there is only a mapId present, the maps must be activated beforehand using instant actions enableMap or initPosition.
- As it should also be possible to activate maps on the vehicle, there is the option of manual activation. This is currently not automatically possible —> We have to discuss this next meeting.
- Current mapStatus that are possible are enabled, for all maps that the vehicle can use, and disabled, for all
* Added json-schema for maps in vehicle state schema.
* Complete rework of Figures in the document, renaming them for single point of truth, providing an uml diagram where possible and updating links in text.
---------

Co-authored-by: Constantin Enke <constantin.enke@kit.edu>