VDA 5050 Zones - Adding Support for freely navigating mobile robots (#362)
A short overview on the introduced concepts and relevant changes to the VDA 5050:

Mobile robots share their planned movement using their state and visualization topic. They share an intermediatePath, which represents the estimated time of arrival at closer waypoints and a plannedPath, which represents a longer path within the robot's currently active order.
There are two categories of zones introduced, contour-based and kinematic center-based zones. This distinction is based on the different conditions for when the vehicle is considered to be entering and exiting zones.
Contour-based zones are 'BLOCKED', 'LINE_GUIDED, RELEASE', 'COORDINATED_REPLANNING', 'SPEED_LIMIT' and 'ACTION' zones.
Kinematic center-based zones are 'PRIORITY', 'PENALTY', 'DIRECTED' and 'BIDIRECTED' zones.
Zones are transmitted to the mobile robot in form of zone sets, that can be either send via a topic or, if the set sizes are to large, via download.
Both a zone and a zone set are well-defined JSON objects. The corresponding json-schemas were added.
To control the traffic, 'RELEASE' and 'COORDINATED_REPLANNING' zones require requests from the mobile robot to the master control, a concept, that is new to the guideline. Requests are stated in the vehicles state, responses by master control are communicated on the new response topic. Definitions, schemas and descriptive figures were added.
Zones can be stacked on top of each other. A conflict matrix is introduced to show the expected interaction between every single one of them.
In the Errors chapter, new issue levels in addition to 'WARNING' and 'FATAL' were introduced, due to requirements of freely navigating vehicles. 'URGENT' was introduced for issues that do not interfere with current order execution but require immediate attention and 'CRITICAL' for errors that cause a mobile robot being unable to continue its current order but it is still able to receive new orders (no user intervention required).